move button into common folder in src
